+++ integration/spark/app/src/main/java/io/openlineage/spark/agent/lifecycle/RddExecutionContext.java
@@ -19,6 +19,7 @@ import io.openlineage.spark.agent.facets.builder.SparkPropertyFacetBuilder;
 import io.openlineage.spark.agent.util.PathUtils;
 import io.openlineage.spark.agent.util.PlanUtils;
 import io.openlineage.spark.agent.util.ScalaConversionUtils;
+import io.openlineage.spark.agent.util.StreamingContextUtils;
 import java.io.IOException;
 import java.lang.reflect.Field;
 import java.net.URI;
@@ -56,15 +57,21 @@ import scala.collection.Seq;
 
 @Slf4j
 class RddExecutionContext implements ExecutionContext {
+  private static final String SPARK_PROCESSING_TYPE_BATCH = "BATCH";
+  private static final String SPARK_PROCESSING_TYPE_STREAMING = "STREAMING";
+  private static final String SPARK_JOB_TYPE = "RDD_JOB";
+
   private final EventEmitter eventEmitter;
   private final Optional<SparkContext> sparkContextOption;
   private final UUID runId = UUID.randomUUID();
   private List<URI> inputs = Collections.emptyList();
   private List<URI> outputs = Collections.emptyList();
   private String jobSuffix;
+  private OpenLineage openLineage;
 
   public RddExecutionContext(EventEmitter eventEmitter) {
     this.eventEmitter = eventEmitter;
+    this.openLineage = new OpenLineage(Versions.OPEN_LINEAGE_PRODUCER_URI);
     Option<SparkContext> activeSessionOption = SparkContext$.MODULE$.getActive();
     sparkContextOption =
         activeSessionOption.isDefined() ? Optional.of(activeSessionOption.get()) : Optional.empty();
@@ -197,14 +204,19 @@ class RddExecutionContext implements ExecutionContext {
       log.info("Output RDDs are empty: skipping sending OpenLineage event");
       return;
     }
-    OpenLineage ol = new OpenLineage(Versions.OPEN_LINEAGE_PRODUCER_URI);
     OpenLineage.RunEvent event =
-        ol.newRunEventBuilder()
+        openLineage
+            .newRunEventBuilder()
             .eventTime(toZonedTime(jobStart.time()))
             .eventType(OpenLineage.RunEvent.EventType.START)
             .inputs(buildInputs(inputs))
             .outputs(buildOutputs(outputs))
-            .run(ol.newRunBuilder().runId(runId).facets(buildRunFacets(null, ol, jobStart)).build())
+            .run(
+                openLineage
+                    .newRunBuilder()
+                    .runId(runId)
+                    .facets(buildRunFacets(null, jobStart))
+                    .build())
             .job(buildJob(jobStart.jobId()))
             .build();
 
@@ -222,18 +234,18 @@ class RddExecutionContext implements ExecutionContext {
       log.info("Output RDDs are empty: skipping sending OpenLineage event");
       return;
     }
-    OpenLineage ol = new OpenLineage(Versions.OPEN_LINEAGE_PRODUCER_URI);
-
     OpenLineage.RunEvent event =
-        ol.newRunEventBuilder()
+        openLineage
+            .newRunEventBuilder()
             .eventTime(toZonedTime(jobEnd.time()))
             .eventType(getEventType(jobEnd.jobResult()))
             .inputs(buildInputs(inputs))
             .outputs(buildOutputs(outputs))
             .run(
-                ol.newRunBuilder()
+                openLineage
+                    .newRunBuilder()
                     .runId(runId)
-                    .facets(buildRunFacets(buildJobErrorFacet(jobEnd.jobResult()), ol, jobEnd))
+                    .facets(buildRunFacets(buildJobErrorFacet(jobEnd.jobResult()), jobEnd))
                     .build())
             .job(buildJob(jobEnd.jobId()))
             .build();
@@ -242,26 +254,25 @@ class RddExecutionContext implements ExecutionContext {
     eventEmitter.emit(event);
   }
 
-  protected OpenLineage.RunFacets buildRunFacets(
-      ErrorFacet jobError, OpenLineage ol, SparkListenerEvent event) {
-    OpenLineage.RunFacetsBuilder runFacetsBuilder = ol.newRunFacetsBuilder();
+  protected OpenLineage.RunFacets buildRunFacets(ErrorFacet jobError, SparkListenerEvent event) {
+    OpenLineage.RunFacetsBuilder runFacetsBuilder = openLineage.newRunFacetsBuilder();
     runFacetsBuilder.parent(buildApplicationParentFacet());
     if (jobError != null) {
       runFacetsBuilder.put("spark.exception", jobError);
     }
 
-    addProcessingEventFacet(runFacetsBuilder, ol);
+    addProcessingEventFacet(runFacetsBuilder);
     addSparkPropertyFacet(runFacetsBuilder, event);
     addSparkJobDetailsFacet(runFacetsBuilder, event);
 
     return runFacetsBuilder.build();
   }
 
-  private void addProcessingEventFacet(OpenLineage.RunFacetsBuilder b0, OpenLineage ol) {
+  private void addProcessingEventFacet(OpenLineage.RunFacetsBuilder b0) {
     sparkContextOption.ifPresent(
         context -> {
           OpenLineage.ProcessingEngineRunFacet facet =
-              new SparkProcessingEngineRunFacetBuilderDelegate(ol, context).buildFacet();
+              new SparkProcessingEngineRunFacetBuilderDelegate(openLineage, context).buildFacet();
           b0.processing_engine(facet);
         });
   }
@@ -299,9 +310,25 @@ class RddExecutionContext implements ExecutionContext {
             .getOverriddenAppName()
             .orElse(sparkContextOption.map(SparkContext::appName).orElse("unknown"));
     String jobName = name + "." + suffix;
-    return new OpenLineage.JobBuilder()
+
+    return openLineage
+        .newJobBuilder()
         .namespace(eventEmitter.getJobNamespace())
         .name(jobName.replaceAll(CAMEL_TO_SNAKE_CASE, "_$1").toLowerCase(Locale.ROOT))
+        .facets(
+            openLineage
+                .newJobFacetsBuilder()
+                .jobType(
+                    openLineage
+                        .newJobTypeJobFacetBuilder()
+                        .jobType(SPARK_JOB_TYPE)
+                        .processingType(
+                            StreamingContextUtils.hasActiveStreamingContext()
+                                ? SPARK_PROCESSING_TYPE_STREAMING
+                                : SPARK_PROCESSING_TYPE_BATCH)
+                        .integration("SPARK")
+                        .build())
+                .build())
         .build();
   }
 
@@ -311,7 +338,8 @@ class RddExecutionContext implements ExecutionContext {
 
   protected OpenLineage.InputDataset buildInputDataset(URI uri) {
     DatasetIdentifier di = PathUtils.fromURI(uri);
-    return new OpenLineage.InputDatasetBuilder()
+    return openLineage
+        .newInputDatasetBuilder()
         .name(di.getName())
         .namespace(di.getNamespace())
         .build();
@@ -319,7 +347,8 @@ class RddExecutionContext implements ExecutionContext {
 
   protected OpenLineage.OutputDataset buildOutputDataset(URI uri) {
     DatasetIdentifier di = PathUtils.fromURI(uri);
-    return new OpenLineage.OutputDatasetBuilder()
+    return openLineage
+        .newOutputDatasetBuilder()
         .name(di.getName())
         .namespace(di.getNamespace())
         .build();